Move plugin declaration to dashboard.plugin
